Show campaign owner and fix delete flow

+++ apps/api/src/app/controllers/pools/collaborators/delete.controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
 import { param } from 'express-validator';
-import { ForbiddenError } from '@thxnetwork/api/util/errors';
+import { ForbiddenError, NotFoundError } from '@thxnetwork/api/util/errors';
 import PoolService from '@thxnetwork/api/services/PoolService';
 import { Collaborator } from '@thxnetwork/api/models/Collaborator';
 
@@ -10,6 +10,7 @@ export const controller = async (req: Request, res: Response) => {
     // #swagger.tags = ['Pools']
     const pool = await PoolService.getById(req.params.id);
     const collaborator = await Collaborator.findOne({ poolId: pool._id, uuid: req.params.uuid });
+    if (!collaborator) throw new NotFoundError('Could not find collaborator');
     if (collaborator.sub === pool.sub) throw new ForbiddenError('Can not remove campaign owner');
 
     await collaborator.deleteOne();
